+++ packages/tolk/test/unit/engine/test_jit.ml
@@ -99,6 +99,7 @@ let call_info name : U.call_info =
     name;
     precompile = false;
     precompile_backward = false;
+    dtype = Dtype.void;
     aux = None;
   }