+++ packages/tolk/test/unit/test_engine_schedule.ml
@@ -12,6 +12,7 @@ let call_info : U.call_info =
     name = None;
     precompile = false;
     precompile_backward = false;
+    dtype = Dtype.void;
     aux = None;
   }